+++ terralab/utils.py
@@ -3,25 +3,15 @@
 import json
 import logging
 import os
-import typing
+import requests
+from tqdm import tqdm
 
 from functools import wraps
-from pydantic import BaseModel
 
 from teaspoons_client import ApiException
 
 
 LOGGER = logging.getLogger(__name__)
-
-
-def _pretty_print(obj: BaseModel):
-    """
-    Prints a pydantic model in a pretty format to the console
-    """
-    try:
-        LOGGER.info(json.dumps(obj.model_dump(), indent=4))
-    except Exception:
-        LOGGER.error(obj)
 
 
 def handle_api_exceptions(func):
@@ -40,7 +30,7 @@ def handle_api_exceptions(func):
     return wrapper
 
 
-def process_json_to_dict(json_data) -> typing.Optional[dict]:
+def process_json_to_dict(json_data) -> dict:
     """Process the given JSON string and return a dictionary.
     Returns None if the input string is not able to be parsed to a dictionary."""
     try:
@@ -51,9 +41,35 @@ def process_json_to_dict(json_data) -> typing.Optional[dict]:
         return data
     except (json.JSONDecodeError, TypeError):
         LOGGER.error("Input string not parsable to a dictionary.")
-        return None
+        exit(1)
 
 
 def is_valid_local_file(local_file_path: str) -> bool:
     """Validate that the provided local file path exists."""
     return True if os.path.exists(local_file_path) else False
+
+
+def upload_file_with_signed_url(local_file_path, signed_url):
+    """Uploads a local file using a signed URL"""
+    try:
+        with open(local_file_path, "rb") as in_file:
+            total_bytes = os.fstat(in_file.fileno()).st_size
+            with tqdm.wrapattr(
+                in_file,
+                "read",
+                total=total_bytes,
+                miniters=1,
+                desc="Upload progress",
+                bar_format="{desc}: {percentage:3.0f}%|{bar}| {n_fmt}/{total_fmt} [elapsed: {elapsed} ETA: {remaining} ({rate_fmt}{postfix})]",
+            ) as file_obj:
+                response = requests.request(
+                    method="PUT",
+                    url=signed_url,
+                    data=file_obj,
+                    headers={"Content-Type": "application/octet-stream"},
+                )
+                response.raise_for_status()
+        LOGGER.info(f"\nFile `{local_file_path}` upload complete\n")
+    except Exception as e:
+        LOGGER.error(f"Error uploading file: {e}")
+        exit(1)
