+++ tests/commands/test_pipeline_runs_commands.py
@@ -4,7 +4,7 @@ import logging
 import uuid
 from click.testing import CliRunner
 from mockito import when
-from terralab.commands import submit_commands
+from terralab.commands import pipeline_runs_commands
 from tests.utils_for_tests import capture_logs
 
 LOGGER = logging.getLogger(__name__)
@@ -19,19 +19,19 @@ def test_submit(capture_logs):
 
     test_job_id = uuid.uuid4()
 
-    when(submit_commands).process_json_to_dict(test_inputs_dict_str).thenReturn(
+    when(pipeline_runs_commands).process_json_to_dict(test_inputs_dict_str).thenReturn(
         test_inputs_dict
     )
-    when(submit_commands.pipelines_logic).validate_pipeline_inputs(
+    when(pipeline_runs_commands.pipelines_logic).validate_pipeline_inputs(
         test_pipeline_name, test_inputs_dict
     )  # do nothing
 
-    when(submit_commands.submit_logic).prepare_upload_start_pipeline_run(
+    when(pipeline_runs_commands.pipeline_runs_logic).prepare_upload_start_pipeline_run(
         test_pipeline_name, 0, test_inputs_dict, ""
     ).thenReturn(test_job_id)
 
     result = runner.invoke(
-        submit_commands.submit,
+        pipeline_runs_commands.submit,
         [test_pipeline_name, "--inputs", test_inputs_dict_str],
     )
