+++ tests/test_utils.py
@@ -3,6 +3,7 @@
 import os
 import pytest
 import tempfile
+from requests.exceptions import HTTPError
 
 from mockito import mock, when
 from terralab import utils
@@ -24,7 +25,12 @@ process_json_testdata = [
 
 @pytest.mark.parametrize("input,expected_output", process_json_testdata)
 def test_process_json_to_dict(input, expected_output):
-    assert utils.process_json_to_dict(input) == expected_output
+    if expected_output is None:
+        # failure
+        with pytest.raises(SystemExit):
+            utils.process_json_to_dict(input)
+    else:
+        assert utils.process_json_to_dict(input) == expected_output
 
 
 def test_is_valid_local_file():
@@ -41,3 +47,53 @@ def test_is_valid_local_file():
 
     # nonexistent file returns False
     assert not (utils.is_valid_local_file("not a file"))
+
+
+def test_upload_file_with_signed_url_success(capture_logs):
+    # Arrange
+    test_signed_url = "signed_url"
+
+    with tempfile.TemporaryDirectory() as tmpdirname:
+        test_local_file_name = "temp_file"
+        test_local_file_path = os.path.join(tmpdirname, test_local_file_name)
+
+        # write the file locally
+        with open(file=test_local_file_path, mode="w") as blob_file:
+            blob_file.write("Hello, World!")
+
+        mock_response = mock()
+        when(mock_response).raise_for_status()  # do nothing
+
+        when(utils.requests).request(...).thenReturn(mock_response)
+
+        # Act
+        utils.upload_file_with_signed_url(test_local_file_path, test_signed_url)
+
+        # Assert
+        assert f"File `{test_local_file_path}` upload complete" in capture_logs.text
+
+
+def test_upload_file_with_signed_url_failed(capture_logs):
+    # Arrange
+    test_signed_url = "signed_url"
+
+    with tempfile.TemporaryDirectory() as tmpdirname:
+        test_local_file_name = "temp_file"
+        test_local_file_path = os.path.join(tmpdirname, test_local_file_name)
+
+        # write the file locally
+        with open(file=test_local_file_path, mode="w") as blob_file:
+            blob_file.write("Hello, World!")
+
+        mock_response = mock()
+        when(mock_response).raise_for_status().thenRaise(
+            HTTPError("some message")
+        )  # raise an error
+
+        when(utils.requests).request(...).thenReturn(mock_response)
+
+        # Act
+        with pytest.raises(SystemExit):
+            utils.upload_file_with_signed_url(test_local_file_path, test_signed_url)
+
+        assert "Error uploading file: some message" in capture_logs.text
